fix: start persistent web socket service
