+++ app/src/main/AndroidManifest.xml
@@ -144,15 +144,6 @@
                 <action android:name="com.google.firebase.MESSAGING_EVENT"/>
             </intent-filter>
         </service>
-        <receiver
-                android:name=".ui.debugscreen.StartServiceReceiver"
-                android:enabled="true"
-                android:exported="false">
-            <intent-filter>
-                <action android:name="android.intent.action.BOOT_COMPLETED"/>
-                <action android:name="android.intent.action.MY_PACKAGE_REPLACED"/>
-            </intent-filter>
-        </receiver>
 
         <receiver
                 android:name=".migration.UpdateReceiver"
